AI_workflow/requirements_optimized.txt
# RTX 3050 Advanced Insurance Query Pipeline Requirements
# Optimized dependencies for production deployment

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Document Processing
PyMuPDF==1.23.8
pymupdf4llm==0.0.5

# Vector Database & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0

# LLM Integration
groq==0.30.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Core Python Libraries
Pillow==10.0.1
scikit-learn==1.3.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Development & Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
